issue #699 Fix server error when decimal value is entered as speed

+++ OpenRobertaRobot/src/main/java/de/fhg/iais/roberta/visitor/validate/AbstractProgramValidatorVisitor.java
@@ -469,7 +469,7 @@ public abstract class AbstractProgramValidatorVisitor extends AbstractCollectorV
     private void checkForZeroSpeed(Expr<Void> speed, Action<Void> action) {
         if ( speed.getKind().hasName("NUM_CONST") ) {
             NumConst<Void> speedNumConst = (NumConst<Void>) speed;
-            if ( Integer.valueOf(speedNumConst.getValue()) == 0 ) {
+            if ( Double.valueOf(speedNumConst.getValue()) == 0 ) {
                 action.addInfo(NepoInfo.warning("MOTOR_SPEED_0"));
                 this.warningCount++;
             }
